+++ zarrs/examples/custom_data_type_uint12.rs
@@ -12,7 +12,7 @@ use zarrs::array::{
 use zarrs::metadata::v3::MetadataV3;
 use zarrs::metadata::{Configuration, FillValueMetadata};
 use zarrs::storage::store::MemoryStore;
-use zarrs_data_type::codec_traits::PackBitsDataTypeTraits;
+use zarrs_data_type::codec_traits::packbits::PackBitsDataTypeTraits;
 use zarrs_data_type::{
     DataTypeFillValueError, DataTypeFillValueMetadataError, DataTypePluginV3, DataTypeTraits,
     FillValue,
@@ -106,8 +106,8 @@ impl PackBitsDataTypeTraits for CustomDataTypeUInt12 {
 // Register packbits codec support
 zarrs_data_type::register_data_type_extension_codec!(
     CustomDataTypeUInt12,
-    zarrs_data_type::codec_traits::PackBitsDataTypePlugin,
-    zarrs_data_type::codec_traits::PackBitsDataTypeTraits
+    zarrs_data_type::codec_traits::packbits::PackBitsDataTypePlugin,
+    zarrs_data_type::codec_traits::packbits::PackBitsDataTypeTraits
 );
 
 impl TryFrom<u64> for CustomDataTypeUInt12Element {